setup: resize, add or remove images when config is changed
Update disk images when their size is changed in the config file.
Disk images may be removed by setting their previously-nonzero size to 0,
created by setting their previously-0 size to nonzero, or recreated with
a different size by changing the size in the config file. An image will
remain unchanged if its size is not changed.