setup: support out of order partitions
Ubuntu apparently supplies disk images where the partition order
in the partition table is different than the order on disk (e.g.
the root partition might be partition 1 in the partition table, but
the last one on the disk).
These changes support that, as long as the root partition is the
last partition on the disk.