Support Bookworm Raspberry Pi OS's /boot/firmware directory
The DOS partition of a Raspberry Pi OS install used to be mounted at /boot,
but is mounted at /boot/firmware since Bookworm. Since this is where
TeslaUSB expects the config file to be, and where it puts the setup log and
various marker files, make /teslausb a symlink to either /boot/firmware (for
Bookworm-based Raspberry Pi OS) or /boot (for older Raspberry Pi OS and other
distributions), and access all the files through this symlink.