Change AP mode dhcp range to match Network manager
When configuring an access point with Network Manager, it uses a dhcp
range of x.x.x.10-254, with no apparent way to change that. Prior to
this change, TeslaUSB would use a dhcp range of 100-150, but this
change changes that to 10-254, so that both methods of configuring an
access point use the same range.